+++ src/main/java/com/mini/buting/api/auth/service/AuthTokenService.java
@@ -1,8 +1,11 @@
 package com.mini.buting.api.auth.service;
 
+import com.mini.buting.global.security.principal.AuthUser;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 
 public interface AuthTokenService {
-    void logout(HttpServletRequest request, HttpServletResponse response);
+    public void logout(HttpServletRequest request, HttpServletResponse response);
+
+    public void issueJwt(AuthUser authUser, HttpServletResponse response);
 }
