+++ src/main/java/com/mini/buting/api/auth/service/impl/AuthTokenServiceImpl.java
@@ -4,16 +4,22 @@ import com.mini.buting.api.auth.service.AuthTokenService;
 import com.mini.buting.global.exception.BaseException;
 import com.mini.buting.global.response.BaseResponseStatus;
 import com.mini.buting.global.security.constant.SecurityConstants;
+import com.mini.buting.global.security.principal.AuthUser;
+import com.mini.buting.global.security.property.SecurityProperties;
 import com.mini.buting.global.security.provider.JwtTokenProvider;
 import com.mini.buting.global.security.service.TokenBlacklistService;
+import com.mini.buting.global.security.token.JwtToken;
 import com.mini.buting.global.util.CookieUtils;
 import com.mini.buting.global.util.RedisUtils;
 import io.jsonwebtoken.Claims;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.GrantedAuthority;
 import org.springframework.stereotype.Service;
 import org.springframework.util.StringUtils;
+
+import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
@@ -23,6 +29,7 @@ public class AuthTokenServiceImpl implements AuthTokenService {
     private final RedisUtils redisUtils;
     private final TokenBlacklistService tokenBlacklistService;
     private final CookieUtils cookieUtils;
+    private final SecurityProperties securityProperties;
 
     /**
      * <h3>세션 단위 로그아웃</h3>
@@ -58,5 +65,24 @@ public class AuthTokenServiceImpl implements AuthTokenService {
 
         // RT 쿠키 만료
         cookieUtils.expireCookie(response, SecurityConstants.Token.REFRESH_COOKIE_NAME);
+    }
+
+    @Override
+    public void issueJwt(AuthUser authUser, HttpServletResponse response) {
+        String memberUuid = authUser.getUuid();
+        String authorities = authUser.getAuthorities().stream()
+                .map(GrantedAuthority::getAuthority)
+                .collect(Collectors.joining(","));
+
+        JwtToken jwtToken = jwtTokenProvider.generateTokenSet(memberUuid, authorities);
+
+        // response.addHeader(HttpHeaders.AUTHORIZATION, SecurityConstants.Token.GRANT_TYPE + jwtToken.accessToken());
+
+        cookieUtils.setCookie(
+                response,
+                SecurityConstants.Token.REFRESH_COOKIE_NAME,
+                jwtToken.refreshToken(),
+                (int) securityProperties.jwt().expireTime().refresh().getSeconds()
+        );
     }
 }
