--- src/main/java/com/mini/buting/global/security/service/TokenAuthService.java
@@ -1,8 +1,0 @@
-package com.mini.buting.global.security.service;
-
-import com.mini.buting.global.security.principal.AuthUser;
-import jakarta.servlet.http.HttpServletResponse;
-
-public interface TokenAuthService {
-    public void issueJwt(AuthUser authUser, HttpServletResponse response);
-}
